requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain-google-genai
langchain==0.2.6
langchain-core==0.2.10
langchain-community==0.2.6
pypdf
python-docx
unstructured
chromadb
tiktoken